rearrange utils
